+++ ml_project/tests/constants.py
@@ -1,0 +1,6 @@
+TRAIN_ROWS = 500
+EVAL_ROWS = 200
+DF_PATH = "data/raw/heart.csv"
+CAT_FEATS = ["sex", "cp", "fbs", "restecg", "exang", "slope", "ca", "thal"]
+NUM_FEATS = ["age", "trestbps", "chol", "thalach", "oldpeak"]
+TARGET_COL = "target"
